lib/wjwt/src/wjwt.ts
// SPDX-FileCopyrightText: Copyright 2024, 谢尔盖||↓ and the Ordo.pink contributors
// SPDX-License-Identifier: AGPL-3.0-only

// Ordo.pink is an all-in-one team workspace.
// Copyright (C) 2024  谢尔盖||↓ and the Ordo.pink contributors

// This program is free software: you can redistribute it and/or modify
// it under the terms of the GNU Affero General Public License as published
// by the Free Software Foundation, either version 3 of the License, or
// (at your option) any later version.

// This program is distributed in the hope that it will be useful,
// but WITHOUT ANY WARRANTY; without even the implied warranty of
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
// GNU Affero General Public License for more details.

// You should have received a copy of the GNU Affero General Public License
// along with this program.  If not, see <https://www.gnu.org/licenses/>.

import type { TWJWT } from "./wjwt.types"
import { decode0 } from "./_decode"
import { sign0 } from "./_sign"
import { verify0 } from "./_verify"

export const WJWT: TWJWT = ({ alg, privateKey, publicKey }) => ({
	sign0: sign0({ alg, key: privateKey }),
	verify0: verify0({ alg, key: publicKey }),
	decode0,
})
